ui/graphs: rename chart-data-source.js to chart-data-fetch.js
Context:
- Review feedback: "source" is reserved for belief sourcing in this repo,
  so naming a module after it is misleading.

Change:
- Rename the module and update its two importers.
