+++ flexmeasures/ui/static/js/chart-data-fetch.js
